gateway/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
httpx>=0.24.0
pyyaml>=6.0
